// 01- Reforzamiento/src/interfaces/reqRes.tsx
/* Para tipar el JSON con TypeScript podemos usar la siguiente página: https://app.quicktype.io 
También podemos usar la extension Paste JSON as Code, con la Command Palette que esta en view
o con el shortcut Ctrl+Shift+P y elegimos TypeScript, luego ponemos el nombre del primer nivel
en mi caso le puse ReqResListado*/// Generated by https://quicktype.io

export interface ReqResListado {
    page:        number;
    per_page:    number;
    total:       number;
    total_pages: number;
    data:        Usuario[];
    support:     Support;
}

export interface Usuario {
    id:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar:     string;
}

export interface Support {
    url:  string;
    text: string;
}